+++ server/src/routes/auth/index.ts
@@ -1,8 +1,7 @@
 import { Router } from 'express';
-import google from './google'
 
 const router = Router();
 
-router.use('/google', google);
+router.use('/google', require('./google').default);
 
 export default router;
