+++ server/src/routes/me/tasks.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+const router = Router();
+
+router.get('/', (req, res) => {
+  res.sendStatus(501);
+});
+
+export default router;
