+++ server/src/routes/projects/index.ts
@@ -1,34 +1,11 @@
 import { Router } from 'express';
 import { projectDAO } from '~/core/projects/project.dao';
-import {
-  ProjectInsert,
-  ProjectUpdate,
-  projectValidation,
-} from '~/core/projects/project.model';
+import { ProjectUpdate } from '~/core/projects/project.model';
 import authenticate from '~/middlewares/authenticate';
-import validateBody from '~/middlewares/validateBody';
 
 const router = Router();
 
-router.post('/', authenticate, validateBody(projectValidation), (req, res) => {
-  const data = new ProjectInsert({
-    ...(req.parsedBody as ProjectInsert),
-    user_id: req.accessToken.user_id,
-  });
-
-  projectDAO
-    .create(data)
-    .then(project => {
-      res.status(201).json({
-        success: true,
-        data: { project },
-      });
-    })
-    .catch(err => {
-      console.error(err);
-      res.json({ err });
-    });
-});
+router.use('/:projectId/tasks', require('./tasks').default)
 
 router
   .route('/:projectId')
